requirements.txt
Flask
beautifulsoup4
google-ai-generativelanguage
google-generativeai
requests
pandas
python-dotenv